enhancements on day 15

+++ src/solutions/day15.rs
@@ -66,12 +66,12 @@ pub fn solution_day_15_01(file_path: String, y: isize) -> Option<usize> {
     let x_bound = (
         sensors
             .iter()
-            .map(|s| s.coordinates.0 - (s.radius as isize))
+            .map(|s| s.coordinates.0 - s.radius)
             .min()
             .unwrap(),
         sensors
             .iter()
-            .map(|s| s.coordinates.0 + (s.radius as isize))
+            .map(|s| s.coordinates.0 + s.radius)
             .max()
             .unwrap(),
     );
@@ -86,8 +86,8 @@ pub fn solution_day_15_01(file_path: String, y: isize) -> Option<usize> {
 pub fn solution_day_15_02(file_path: String, bound: isize) -> Option<isize> {
     let sensors = parse_input(file_path);
     sensors.iter().find_map(|s| {
-        ((s.coordinates.0 - (s.radius as isize) - 1).max(0)..=s.coordinates.0.min(bound))
-            .zip(s.coordinates.1..=(s.coordinates.1 + (s.radius as isize)).min(bound))
+        ((s.coordinates.0 - s.radius - 1).max(0)..=s.coordinates.0.min(bound))
+            .zip(s.coordinates.1..=(s.coordinates.1 + s.radius).min(bound))
             .find_map(|coords| {
                 sensors
                     .iter()
